Benchmark the wavetable graph, and correct a claim it did not support
The graph repaints its region in situ at 0.898 ms per frame while the scan
moves, comfortably inside a 16.7 ms budget. Measured through the editor rather
than on the component alone, because a repaint redraws everything beneath it -
the same measurement taken in isolation earlier in this project read 0.02 ms
where the real figure was 27.6 ms.

The cached surface, however, does not pay for itself. Removing it entirely
measures 0.819 ms against the cache's 0.898 ms: both are dominated by the editor
beneath rather than by forty polylines of 192 points, and the cache is, if
anything, marginally the slower of the two. The comment claiming that
regenerating the geometry every repaint "is what makes a wavetable display
expensive" was an assumption, and it was wrong at this size.

Kept anyway, because the display size is a parameter and a 256-frame table is
six times the geometry - but the header now says that is why, and says to delete
it if that stops being true. Recorded rather than quietly left in place, since
the original justification is the sort that would otherwise be read later as
having been measured.

780 tests passing, 0 allocations per block.

+++ Source/UI/WavetableGraph.cpp
@@ -176,7 +176,9 @@ void WavetableGraph::rebuildSurface()
     // Back to front, so nearer frames overlap the ones behind them.
     for (int f = frameCount - 1; f >= 0; --f)
     {
-        const auto depth = frameCount > 1 ? static_cast<float>(f) / (frameCount - 1) : 0.0f;
+        const auto depth = frameCount > 1
+                             ? static_cast<float>(f) / static_cast<float>(frameCount - 1)
+                             : 0.0f;
         const auto& points = display.frames[static_cast<std::size_t>(f)];
         if (points.size() < 2)
         {
@@ -190,7 +192,8 @@ void WavetableGraph::rebuildSurface()
         juce::Path path;
         for (std::size_t i = 0; i < points.size(); ++i)
         {
-            const auto x = originX + width * static_cast<float>(i) / (points.size() - 1);
+            const auto x = originX + width * static_cast<float>(i)
+                                       / static_cast<float>(points.size() - 1);
             const auto y = originY - points[i] * amplitude;
             if (i == 0) { path.startNewSubPath(x, y); }
             else { path.lineTo(x, y); }
